move action plugin to system/ and replace it by a symlink
